Put ddah deletion confirmation modal into separate file.
